Color Detector & Tracker/requirements.txt
numpy==1.18.1
opencv_contrib_python==4.2.0.34
requests==2.22.0
pandas==0.25.3
Pillow==8.1.0
